Proposed changes: Linked to issue: https://github.com/RasaHQ/rasa/issues/5574
The default action 'action_session_start' wasn't
able to receive the metadata from the message sent,
RASA when initiating a new session it sends a blank
tracker with no information whatsoever, ignoring
the metadata into the message.

Adjusted to recieve the event sessionStart from tracker events
and with it the metadata from the message.